Update service for Kubernetes client-go v0.18.x
The Go team recommends using a context.Context as the first parameter
to provide timeouts, deadlines and cancellation functions.

The previous versions of the Kubernetes libraries did not use contexts.
The new versions (v0.18+) include contexts, introducing a backwards
incompatibility.

This commit updates to v0.18.02. It changes the versions in go.mod and
adds contexts to the orch package.