Add support for GCP secret during pod creation
This commit adds support for mounting a secret volume, allowing driver
pods to use a GCP service account. Developers should use the service
account to write benchmark results and artifacts to non-volatile
storage.

The developer must manually add a Kubernetes secret with a security key
for the service account. Then, they specify the name of the secret in an
environment variable: $GCP_KEY_SECRET. Finally, the system will detect
the presence of the environment variable and attempt to mount the secret
into the driver component's container.

The previous version of the orch package introduced a SpecBuilder type.
Executors used this type to create Pod instances. This type allowed the
Pod spec to be broken into multiple methods. Each method focused on one
aspect of the pod, such as environment variables. Unfortunately, volumes
requiring sharing data across these methods. Therefore, this commit
replaces the SpecBuilder type with one function `makePod` and its helper
`scenarioJSON`.

Finally, this commit intentionally lacks automated testing of the new
functions and secret business logic. The team plans to move to a custom
controller, causing this logic to be split into a new package. During
the split, tests will be added using the ginkgo framework.