Fix clone init container idempotency
According to the Kubernetes documentation for init containers,
developers should assume that they may be restarted or rerun at any
time. They are expected to be idempotent.

The clone container does not address the situation where the
/src/workspace directory exists and is not empty. This change adds a
command to remove the pre-existing contents of the /src/workspace
directory. This removal ensures that the output of the container is
equivalent, regardless of pre-existing runs.