Add name field to LoadTest Component
In order to distinguish amongst similar components in a list, this
change adds a name field to the LoadTest. For example, a user may
specify a list of client components. If one component fails but the
other is running, it becomes important to distinguish the failing one.

Names must be unique within a LoadTest. If a user does not specify a
name, the system should dynamically assign one. This logic will be added
in a later commit.