Add errors package for sane error wrapping
Consider that the `error` type is an interface. A struct can implement
the `error` interface. By wrapping a error-implementing struct with
`fmt.Errorf`, the data structure is lost. It is converted to a string.

The github.com/pkg/errors package provides a safe way to wrap and unwrap
errors losslessly. In other words, it allows additional information to
be added to the error while retaining its type.