Fix defaults variable shadowing package
The main function declares a variable named `defaults`. This is also the
name of the package with the type of that variable. This causes the
package to be shadowed. Variable shadowing can lead to subtle issues.

This change renames the variable to `defaultOptions`, so the word
"defaults" refers to the package alone.