Switch pkg/defaults to whitebox testing
Ginkgo encourages blackbox testing, which means tests live in a
separate package from the implementation. It generates test files with a
package name that adds a `_test` suffix. pkg/defaults uses whitebox
testing, allowing the tests to live in the same package.

Unfortunately, the `_test` suffix was not removed from the package name
in the file defaults_suite_test.go. This change removes it, ensuring
that all files use the defaults package.