Fix typos in language defaults
The `Languages` field on the Defaults object declared a json tag of
"buildImages" instead of "languages". This caused it to be parsed
incorrectly.

In addition, the example config referenced a `Name` field which does not
exist on the LanguageDefault struct. It was renamed to `Language`.

This change corrects both mistakes.