Add comments for imageMap and its methods
The initial version did not contain any documentation for the imageMap
type and its methods. These are designed for internal use and are not
exported from pkg/defaults.

However, documentation may be helpful for onboarding new contributors.
This change adds documentation comments to the imageMap structure, its
constructor function and its methods.