Set unique name for pods
The name of a component must be relatively unique. For example, two
client components in the same load test cannot share the same name.
However, the names of client components can be shared across different
load tests.

This commit sets the name of a component's pod to be a combination of
the name of the load test, the role of the pod and the component name.
This name is unique across the namespace since the following criteria
are satisfied:

 - Kubernetes rejects load tests with the same name

 - Roles distinguish between server, client and driver components

 - Component names uniquely distinguish between two components with the
   same role in the same test