Generate unique component names when unspecified (#19)
Each driver, server and client component must have a relatively unique
name. This means their name cannot be identical to another component
with the same role in the same load test. Previously, this name was
manually provided.

This change generates unique names for components where names are not
manually set. To ensure relative uniqueness, it uses Google's UUID
library to generate a token.