Build driver code in driver Dockerfile (#22)
The previous version of the driver image prepared an environment for
running the driver but did not include a compiled version of the driver
in the container image. The driver code would be built at runtime.
However, this would result in slow Java and Go tests, since the driver
takes much longer to build than Java or Go's workers.

This change compiles the latest driver from the master branch of
grpc/grpc whenever the container image is rebuilt. This allows a driver
to start almost instantly!