Set default namespace on load tests and their pods
The previous version of the code did not specify a namespace when
creating pods for load tests. This led to a panic, because the default
namespace is not inferred.

This change explicitly sets the namespace of pods to match the
namespace of the load test. Though unlikely to be required, it also
ensures that the load test lives in the default namespace if another has
not been specified.