Remove unnecessary logs in Go GitHub Actions
The Go GitHub Actions workflow logged every package (including third
party ones) as it built or vetted them. This led to errors that were
hidden in a large scroll-back of package names. This change removes the
unnecessary logs in the build and vet steps. No logs are now indicative
of success.