Fix iterator reference issue in defaults package
The defaults package contains an internal type called imageMap. This
type is used to efficiently lookup default images for a language. During
the creation of its internal map, the code relied on the address of an
iterator. This led to the map being populated with only the last value.
This change fixes this issue.