Switch to new and pointers for creating resources
The previous version of the controller declared literals for a pod list,
node list, load test list and a load test. The methods used by the
controller require a pointer receiver.

Therefore, this commit changes these variables to be pointers, allocated
and set to their zero values with the built-in new function.