Add worker address file parsing to driver
The driver needs a comma-delimited list of IP addresses and port numbers
for workers. It receives this list from the `$QPS_WORKERS` environment
variable.

This change makes it possible to set a `$QPS_WORKERS_FILE` environment
variable on the driver. When present, the driver container copies the
contents of the file into the `$QPS_WORKERS` environment variable before
invoking the driver executable. This allows the driver to learn these
addresses through a shared volume.