Fix issue with overridden workspace in the driver
The clone, build and run pipeline shares files in the /src/workspace
directory over a volume. The driver includes the code and executable in
the image. However, the volume is always mounted and initially an empty
directory. This caused all files to be deleted before the driver runs.

This change places the included code and executable in a path separate
from the /src/workspace mount. This will allow the driver to work out of
the box or from a version cloned and built at runtime.