Setup the ready init container for a driver
This change configures an init container that waits for workers to
become ready before allowing the driver to start. This init container
also writes the IP addresses and ports of available workers to a file.
This change shares that file over a volume with the driver's run
container.

The default init container with this functionality was added in
grpc/test-infra#27.