Fix plurality for $SCENARIOS_FILE env variable
The driver container looks for a `$SCENARIOS_FILE` environment variable
which contains the path to a file with a list of scenarios to run. This
was added in #22. By mistake, the controller set a singular version,
$SCENARIO_FILE, which prevented the driver from finding and running
scenarios. This change corrects the controller to use the plural
version, `$SCENARIOS_FILE`. And, this resolves the issue.