Remove listing of nodes
Listing nodes will be required to perform gang scheduling in the future,
but the system is not configured with the correct RBAC roles to access
nodes. These roles will need to be added in order to list nodes.

Since gang scheduling is not an immediate priority, this commit comments
out logic to list nodes. This fixes an issue where load tests are not
reconciled due to incorrect roles.