Add State and determine State of a container
This change adds a State enum with three possibilities: Pending,
Succeeded and Errored. This type will be used to convey the health of a
container and pod.

As a first step, this change also adds a StateForContainerStatus
function. This function returns the State and exit code (if applicable)
for a container status.