Add optional package with convenience functions
Kubernetes often uses pointers to convey optionality. This is true for
primitive types, like integers and strings. Since declarations and
taking the address of a literal cannot occur on the same line in Go,
this commit adds an optional package. This package provides convenient
functions to declare and take the address of simple values in one line.