Prevent load test restarts
The controller attempts to reconcile a load test whenever the state of
the world does not match its spec. This led to restarting the test when
one of its pods were cleaned up.

This change prevents the restart by checking the previous state of the
test. If the test was previously marked as a success, failure or error,
the controller now does no work.