Fix issue where pending containers are hidden
This commit fixes an issue with checking the status of a pod with
multiple containers. The check was dependent on order, meaning the check
may incorrectly classify a pod based on the order in which it encounters
the status of its containers.

For example, a pod may have 2 containers. The first container may be
marked as Pending. The second container may be marked as Succeeded.
Because the second container was successful, it masked the Pending
status of the first. The pod was marked Successful.

This change resolves this problem by setting the pod status to be the
first container status. Then, the remaining containers are iterated
through. If any have a Pending or Failed state, it will bubble up to
become the status of the pod.