Add bazel cache volume
The driver and bazel-built workers were failing, because the official
bazel image has a complex output behavior. Bazel compiles into a bazel
cache directory. Then, it creates links in the bazel-bin and bazel-out
directories to the executables in the bazel cache.

This link structure does not allow the build init container to deliver
an executable to the run container. The run container lacks the bazel
cache available to the build init container. Therefore, the bazel-bin
and bazel-out directories link to nothing accessible and the executables
are not present.

To resolve this issue, this commit introduces a bazel cache volume on
all pods. It mounts the volume in the build init container and run
container. When languages do not use bazel, the mounted directory is
empty and has no effect. When languages use bazel, the mounted directory
shares the bazel cache. This allows the bazel-bin and bazel-out links to
resolve, making executables accessible to the run container.