Raise ready init container timeout to 15 minutes
The ready init container was set to timeout after 5 minutes, marking the
test as erroring. Unfortunately, this is not enough time to build the
C-based gRPC through bazel. On 2020-10-07, the C++ example scenario took
7 minutes and 8 seconds to compile off grpc/grpc at master and run the
test. I would expect that a slow build may exceed 10 minutes. For these
reasons, this commit raises the timeout of the ready init container
process to 15 minutes.