Rename go.yml to go.yaml and reformat
The GitHub workflow to build and test the Go code in the project was
defined in a go.yml file. The extension `.yml` is preferred by GitHub;
however, the project uses `.yaml`. This commit renames it to be
consistent with the rest of the project.

In addition, this commit reformats the YAML using the prettier tool. See
https://prettier.io.