Remove container lookups by index
Each pod has a list of containers and init containers. In the past, the
run container was quickly accessed using index 0. It was the first and
only container for pods created by the load test controller. This lookup
was not clean, since the introduction of additional containers may lead
to bugs.

This change introduces a ContainerForName function in the kubehelpers
package. This function iterates over a list of containers. It returns a
pointer to the one with a name that matches a given string. This
function replaces the index lookups throughout the non-prototype code.