Add flag for timeouts to reconcile
This change removes a magic value and unused constant that were both
intended to set a timeout for a single invocation of the Reconcile
method. It replaces them by adding a `-reconciliation-timeout` flag on
the manager. This flag's value is provided to the LoadTestReconciler
upon initialization.

In addition, this commit sets the flag to 2 minutes for the deployment
of the manager and its controller. This number is consistent with the
previous magic value.