Optimize branch checkouts in clone init container
This change optimizes the checkout of a branch or tag for the clone init
container. Previously, repositories were cloned off the master branch.
This commit switches to cloning a specific, single-branch only.

Unfortunately, this breaks checking out a commit hash, because a tree
must be downloaded for git to resolve a specific commit.