Remove unnecessary checkouts when cloning
The clone docker image is used as an init container which fetches the
code at a specific commit, tag or branch. Previously, it used the
git-clone command, which clones the master branch by default. Developers
wanted to avoid the unnecessary checkout when they are testing off
another branch or tag.

The git-clone command does provide the --branch and --single-branch
flags, but these options are incompatible with commits due to a
server-side setting that GitHub has not enabled.

This commit achieves similar behavior, but it uses an empty git
repository, fetches a remote and resets to a specific branch, tag or
commit. This avoids the unnecessary checkout.