Remove gRPC version from driver image tag
The driver-image make target builds the driver from grpc/grpc at the
version specified by the $DRIVER_VERSION environment variable. It tags
the image with the $TEST_INFRA_VERSION environment variable and the
$DRIVER_VERSION.

This does not match the configuration template
config/defaults_template.yaml. This change removes the $DRIVER_VERSION
from the tag.