Decide when to requeue based on status difference
This commit checks previous status of a loadtest and compare that
to its current status. If the previous status of the loadtest does
not have a start time but after status update, the loadtest now
have a start time, we will then requeue the loadtest with timeout.
Similar idea goes to stop time too.