Embed Scenarios in load tests
This change adds support for embedding a JSON string with the Scenarios
message on a load test. Under the hood, the controller will create and
manage the ConfigMap. It also removes support for referencing a manually
created ConfigMap. In accordance with this change, this commit also
modifies all of the examples to support this new style.

Finally, it comments out some tests with TODO markers. In a subsequent
commit, the tests will be improved and cover this embedding thoroughly.