Change scenarios to use official names
This commit changes the Go, Java and Python scenarios to use the
official names. Previously, they were called
"<language>_example_scenario", so it matched the ConfigMap names. In
addition, it changes the Go scenario to use the secure version. It
previously used the insecure version which was inconsistent with the
other languages.