Add controller RBAC permissions for config maps
In order to get, create and delete ConfigMap resources; these
permissions must be specified in the ClusterRole resource that is
assigned to the controller. This commit adds the markers and generated
code to set these permissions.

This change is required to embed scenarios in a production environment
and serves as a required follow-up to grpc/test-infra#92.