Mark load tests as owners of config maps
Every reconciler declares the relationship between a resource and the
subresources it creates. Despite the LoadTestReconciler creating config
maps in #94, it did not declare ownership of them. This means that
Reconcile was not invoked when the ConfigMap resources changed.

This change denotes the relationship between a LoadTest resource and a
ConfigMap resource. It is a best practice.