Remove components and extract pod creation logic
This commit removes the Component type. The Driver, Server and Client
types shared fields through composition with a Component. However, this
made the code hard to understand. It also made it difficult to add a
field to one of these types but not all of them. By removing the
Component type and composition, the code is simplified and more
future-proof.

In addition, this commit moves the logic for creating pods out of the
controller package. It adds a new podbuilder package, which contains a
PodBuilder type. This type can create pods for Drivers, Servers and
Clients.