Add cleanup_agent as a separate component to send RPC (#89)
* Add cleanup_agent as a separate component to send RPC

This commits add a cleanup_agent as a separate component who watch
events of LoadTest and send quit RPC to stop unwanted processes
running in workers. The cleanup_agent does nothing when the triggering
LoadTest is not terminated. The cleanup_agent actively checks the status of pod 
based on the container status when the triggering LoadTest is terminated,
when the pods belong to triggering LoadTest is not terminated, the 
cleanup_agent send RPC to stop the workers.