Add additional tests for pod failure granularity
The previous version of the e2e tests considered all pods as failing at
the same point in time. This would not show issues in the logic if one
but not all of the pods failed. This commit adds additional test cases
to show the correct behavior when the client, driver or server pods
fail independently.