Fix status updates when duplicate pods are created
Controllers are level-driven and operate with a cache, meaning
invocations of the Reconcile method on the LoadTestReconciler may not
always be able to access up-to-date information. When many invocations
occur in a short timespan, the cache may not reflect the actual state of
the cluster. Since a newly created pod may not be immediately noticed
via the cache, multiple pod creation attempts may occur.

Previously, multiple attempts to create a pod resulted in a Kubernetes
error. The test was considered to have errored. However, often the pods
were created successfully and there was no mechanism to stop the test.
The test could be marked as an error but continue to run successfully.

This change adds checks to prevent test errors when the controller
attempts to create a duplicate pod.