Add controller fields to mock client methods
This commit adds function fields that match the signatures defined in
the client.Client interface to the LoadTestReconciler. During normal
instantiation, it sets the value of these fields to be their
client.Client method equivalents. Finally, it replaces all invocations
of the client.Client methods with invocations of these fields. This
indirection provides an easy way to mock out these functions for unit
testing.