Add server port flag and explanation to example
The C++ workers in grpc/grpc rely on a server to find available ports
when no explicit port is given. Unfortunately, this script relies on
Python's subprocess package. The subprocess package has some
compatibility issues with Docker, creating an issue where the port
server could not start.

To temporarily resolve this problem, this commit explicitly introduces a
--server_port flag in the run section of the server for the C++ example
LoadTest YAML file. Specifying this flag disables the need for a port
server, bypassing the problematic code.