Feature/node image (#108)
* Feature/Node.js build and runtime images

This commit adds and updates files to support Node.js tests. The
support added is sufficient to bring up the workers in the image
and start an example LoadTest through submitting the LoadTest
configuration file. However, the provided Node.js example LoadTest
will not success due to a know issue related to the test itself
other than OSS gRPC benchmark infrastructure. The changes made
through this commit should be enough to support Node.js LoadTest
once the issue is fixed.