Fix godoc for runner package
When documenting a package with many files, the community conventions
create a doc.go file with the package documentation comments. The other
files do not include documentation comments for the package. This allows
it to live in one place. This commit modifies the runner to have this
doc.go file and removes the package documentation comments from the
other files.