Fix segfault when no XML report is generated
The runner can generate JUnit XML reports using the reporter structs.
The reporters log to stdout even when reports are not generated. An
index uniquely identifies a load test in these logs. In the previous
version, the value of this index was determined using the number of
known test cases. Unfortunately, this reference is nil when JUnit
reports are disabled. This led to a crash.

This commit removes this reference. It adds an integer that is manually
incremented to provide a unique index to each test case. This resolves
the issue.