Add timeout for workers (#182)
This commit adds timeout for workers. Due to the complication of
directly pass the command and argument through LoadTest
configuration yaml files, a script is need to start the worker
with timeout. To make the script work, the argument
--driver_port of the run container has to be exposed. Currently
--driver_port is set as an environmental variable and passed to
worker at starting time. The templates are also updated to reflect
the change.