Merge kubebuilder3 to master. (#209)
* ignore empty default pool

* error our when default pool is used but undefined

* prettier code

* Upgrade to kubebuilder 3 (#187)

This change upgrades the project to use Kubebuilder 3 and newer
Kubernetes APIs. For details of all the included changes, please refer
to the Kubebuilder migration docs:

  https://book.kubebuilder.io/migration/v2vsv3.html

* resolve php7 build error, add leases access to controller

* Synchronize `kubebuilder3` branch with master. (#203)

* Update defatuls_template.yaml for php7 (#186)

* Update example tests to include named drivers. (#189)

Examples are generated from https://github.com/grpc/grpc/pull/26501.

* Add timeout for workers (#182)

This commit adds timeout for workers. Due to the complication of
directly pass the command and argument through LoadTest
configuration yaml files, a script is need to start the worker
with timeout. To make the script work, the argument
--driver_port of the run container has to be exposed. Currently
--driver_port is set as an environmental variable and passed to
worker at starting time. The templates are also updated to reflect
the change.

* Fix executable path in Java Loadtest example with prebuilt image (#190)

* Add governance and maintainers files. (#191)

* Update driver timeout (#192)

This commit update where the timeout is applied for driver
process. The timeout was wrapped around the driver
process only, which left other operations such as uploading
results to BigQuery unguarded. The timeout is updated
to wrap around the whole operations to make sure driver
pod will timeout after given time.

* Update README files to include test runner. (#194)

* Update tools README to include test runner.

* Move tool binaries under the tools directory. (#196)

* Feature/add php7_protobuf_c (#195)

The PR wraps timeout on the whole operation of php7 worker. 
Also add support php7_protobuf_c tests.

* Initial fixes for xml report generation. (#198)

* Count errors instead of failures.
* Stop counting warnings as errors.
* Keep `id` attribute but only for `testsuite`, as a count starting from zero.
* Rename package to `xunit`.

* Remove unneeded copy from xml report generation. (#200)

* Ignore empty default pool (#201)

* ignore empty default pool

* error our when default pool is used but undefined

* prettier code

Co-authored-by: Wanlin Du <67486458+wanlin31@users.noreply.github.com>
Co-authored-by: Paulo Castello da Costa <6579971+paulosjca@users.noreply.github.com>

* Solve flakiness (#205)

* Separate deployment of cleanup agent from main process (#202)

This pr separate the deploying cleanup agent from the main deploy process.
This is because the cleanup agent is no longer necessary in terms of killing
the leftover pods.

* solve conflict

* Delete all cleanup agent codes (#204)

* resolve conflict

* Synchronize kubebuilder3 branch with master. (#206)

* Separate deployment of cleanup agent from main process (#202)

This pr separate the deploying cleanup agent from the main deploy process.
This is because the cleanup agent is no longer necessary in terms of killing
the leftover pods.

* solve conflict

* Delete all cleanup agent codes (#204)

* resolve conflict

Co-authored-by: Wanlin Du <67486458+wanlin31@users.noreply.github.com>

* prettier style

* prettier style

* merge master to kubebuilder3 to solve conflicts. (#207)

* Update defatuls_template.yaml for php7 (#186)

* Update example tests to include named drivers. (#189)

Examples are generated from https://github.com/grpc/grpc/pull/26501.

* Add timeout for workers (#182)

This commit adds timeout for workers. Due to the complication of
directly pass the command and argument through LoadTest
configuration yaml files, a script is need to start the worker
with timeout. To make the script work, the argument
--driver_port of the run container has to be exposed. Currently
--driver_port is set as an environmental variable and passed to
worker at starting time. The templates are also updated to reflect
the change.

* Fix executable path in Java Loadtest example with prebuilt image (#190)

* Add governance and maintainers files. (#191)

* Update driver timeout (#192)

This commit update where the timeout is applied for driver
process. The timeout was wrapped around the driver
process only, which left other operations such as uploading
results to BigQuery unguarded. The timeout is updated
to wrap around the whole operations to make sure driver
pod will timeout after given time.

* Update README files to include test runner. (#194)

* Update tools README to include test runner.

* Move tool binaries under the tools directory. (#196)

* Feature/add php7_protobuf_c (#195)

The PR wraps timeout on the whole operation of php7 worker. 
Also add support php7_protobuf_c tests.

* change gpg keyserver name (#197)

* Initial fixes for xml report generation. (#198)

* Count errors instead of failures.
* Stop counting warnings as errors.
* Keep `id` attribute but only for `testsuite`, as a count starting from zero.
* Rename package to `xunit`.

* Remove unneeded copy from xml report generation. (#200)

* ignore empty default pool

* error our when default pool is used but undefined

* prettier code

* Ignore empty default pool (#201)

* ignore empty default pool

* error our when default pool is used but undefined

* prettier code

* Separate deployment of cleanup agent from main process (#202)

This pr separate the deploying cleanup agent from the main deploy process.
This is because the cleanup agent is no longer necessary in terms of killing
the leftover pods.

* Delete all cleanup agent codes (#204)

* Separate deployment of cleanup agent from main process (#202)

This pr separate the deploying cleanup agent from the main deploy process.
This is because the cleanup agent is no longer necessary in terms of killing
the leftover pods.

* solve conflict

* Delete all cleanup agent codes (#204)

* resolve conflict

* prettier style

* prettier style

Co-authored-by: Wanlin Du <67486458+wanlin31@users.noreply.github.com>
Co-authored-by: Paulo Castello da Costa <6579971+paulosjca@users.noreply.github.com>

Co-authored-by: Ben Reed <benreed@google.com>
Co-authored-by: Wanlin Du <67486458+wanlin31@users.noreply.github.com>
Co-authored-by: Paulo Castello da Costa <6579971+paulosjca@users.noreply.github.com>