Get logs from all pods, not just the driver
There was a bit of refactoring in this commit. This was necessary
because now there can be errors writing logs to multiple files. If there
is an error writing one log, it shouldn't impact writing the others. If
there are issues writing multiple files, we want to log exactly what
logs weren't able to be saved and the specific reason for failure.