Marshal and Unmarshal config file of snapshot of xds server (#245)
This PR create a config library to marshal and unmarshal configuration for xDS server. This PR include:

The config library provides:
Marshal and unmarshal configuration json file for xDS server.
Checks to make sure that:
The resource are consistent to be served.
Validate that at least one of the listener has the required name
Validate at least one of the listener have the proper port.
A piece of default configuration json file.
Function to update the endpoints. Will report an error if the actual number of the backend are different from the endpoint configurations provided.
Preliminary documentation for using config library.