Update podbuilder for multiple run container (#273)
This commit updates the podbuilder package to accept multiple run
containers and make sure that always identify the main container
by position. The podbuilder adds the WorkingDir and VolumeMounts
for the main (the first run) container and ENV for all run
containers.Added container name for the first run container to be closer
to the actual situation.