Update structure of Prometheus folder.
This commit makes following changes:
1. Delete namespace.yaml since the prometheus and prometheus
operator are now deployed in the test-infra-system namespace.
2. All other files in the config/prometheus/setup are moved to
config/prometheus. The crds.yaml contains all necessary crds and
are moved to config/prometheus/crds/bases. kustomization.yaml
file is updated for this change.
3. Files originally in config/prometheus are deleted since they
were not used.
4. Makefile targets are updated, the prometheus crds installation
and deployment for prometheus operator and prometheus are separated,
same applied to uninstall and undeploy.