Make collecting Prometheus data optional through annotation
Currently we automatically collect Prometheus data for PSM
related tests, now some regular tests (same scenario as PSM
test) are also part of the PSM tests, we may want to collect
Prometheus data for those regular tests too. The current method
for checking if a rest is a PSM test to check the existence of
xds-server and sidecar containers.
It is hard to apply the same mechanism for PSM related regular
tests,  the simplest way is to add an annotation like
`enablePrometheus:true` to control if we want to use Prometheus
for a given test.